pyproject.toml
[project]
name = "flask-easy-oidc"
version = "0.1.0"
description = "Flask extension for an OIDC client."
dependencies = [
    "click", "flask", "requests", "PyJWT", "cryptography"
]
readme = "README.md"
license = {text = "MIT License"}
keywords = ["oidc"]

[project.urls]
Homepage = "https://github.com/andycaine/flask-easy-oidc"
Documentation = "https://github.com/andycaine/flask-easy-oidc?tab=readme-ov-file"
Repository = "https://github.com/andycaine/flask-easy-oidc.git"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
